Added a custom Project TreeStructureProvider.
The project view shows the unqualified module name for D files ("foo",
not "io.github.intellij.dlanguage.foo"). If the module name and file
name are different, then both are shown, with the file name appended in
gray. In the case of "package.d" files, they are displayed simply as
"package.d", not the package name.