Performance: reduce input latency derived from formatting model.
getSpacing is called many times when the user types the return key to insert a
newline. The call to hasErrorElements only takes a few ms, but since this is an
inner loop it ends up adding nearly 80ms of input latency for a ~500 line file,
enough to make editing feel laggy.  Removing this call significantly improves
the editing experience.

Note that other IntelliJ plugins do not attempt to validate syntax at this
phase, so I think we should be okay to omit the check.

fixes #611