#684 Prevent IDE to hangs when configuring Tools
When the update was done in the thread and that we close the window, the
thread try to update a field to a window that is being deleted. For some
reason this was making the IDE to hang due to a deadlock.
This fix prevent deadlock by ensuring all threads that could modify a
field is stopped before releasing the window