#734, #846: Improve resolve of imports and warnings of undefined symbols (#860)
* Re-enable fix to make gradle able to find tests

* Move resolve test to java

* Convert test resolve files to kotlin

* Add import case and fix it

When a class has an explicit constructor (or destructor) and we import the class, we
don’t want to point to it but only to the class definition.

* Remove broken test

It was exactly the same as ImportNameBindResolve but with succeed =
false and was passing (seems wrong).

* Add test for renamed import and fix

* Change not found imports to something more consistent with other IDEs

Contrary to other elements, no mixin or static if can define module, so
if we don’t find the module, we can be certain that it does not exists.

* Move java module index to kotlin

* Convert DModuleIndex to kotlin

* Add comment to explain to goal of the DModuleIndex

* Add all possible names of a package.d when it has multiple

This appear when the module is in A.B package but has a "module foo.bar"
declaration in it. It appears that both names are valid from current DMD
point of view.

* Add some tests to check for simple imports resolution

* Add tests to check for resolve of import usage

* Don’t generate warning for identifier in version() and in debug()

It’s fine for them to not exists. If they not exists, this mean that the
version/debug is not enabled.

* Don’t generate a warning for name of a renamed import

It’s a declaration of the identifier

* Improve doc comment syntax

* Fix #734 support renamed import bindings

---------

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>