fix issue in DResolveUtil.kt and related test (#891)
* Fix wrong get reference at call in test

This gives the wrong reference making a test wrongly passing instead of
failing. The constructor call is not resolved to the constructor
currently but to his class declaration. The current test was passing
because the current resolve is resolving the Class instead of the
constructor call.

* Fix test

The constructor call should resolve to the constructor if defined
otherwise to the class definition.

---------

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>