Introduce a submodule for the plugin implementation
This allow to break the current circular dependency between dub and the
core plugin. A long term plan is to split the core implementation into
more submodules, implementation that need to remain private (for the
plugin only) and the api that is shared and can be used by other
modules. A more complex architecture (with more isolated submodules)
can be made in the future but the goal of this commit is to be able to
upgrade intellij-gradle plugin.