Revwrite the parser Step 1
Remove the Token (and Token.IdType) abstraction and directly use
IElementType instead. This allow to simplify logic, make things more
clear and actually fixed bugs.

First bug was that register were not recognized due to text of token
being invalid (was containing DlangTokenType.ID instead of the name of
the id (EAX)).
Second bug was that a missing `.type` in parse pragma was here so the comparison was
always returning false.
The last bug was for the parsing of associative arrays were the
initialization was not detected correctly.