Add psiViewer plugin for local dev
This plugin is really useful when developping a language plugin, so it’s
usefull to always having it when starting a local instance. However
don’t install it for CICD as it’s not useful there and the version may
be incompatible with the variety of IDE version it covers.