Redo the resolve algorithms (#987)
* Introduce a PackageReference to resolve packages and modules

Having a dedicated Reference for package simplify the resolve algorithms
by having multiple specialized algorithms instead of one big generic and
complex.

* Use the new PackageReference for IdentifierChain elements

* Introduce a UserDefinedType interface

This will allow to handle all user defined types/aggregates (struct,
enum, class, interface, union) at the same place using this interface.

* Introduce new TypeReference for qualified identifier

* Introduce new implementation to resolve the types

* Switch TypeReference to use the new algorithm

* Use the new TypeReference in QualifiedIdentifier

* Add some resolve test about local definitions

* Fix resolve for local (in block statement) declarations

With this fix something used before being declared won’t resolve, as
expected.

* Fix tests about the resolve of the constructor call

To handle properly the rename, the constructor call must resolve to the
class itself. The goto behavior (and highlight) making think that
constructor call actually resolve to the constructor itself is handled
by anothers mechanisms that will be implemented later.

* Add support for resolve of opaque class

* Add test for resolve of an opaque class usage

* Fix and refactor processDeclaration function

When we process some declaration (like struct, or class) we just want to
process the declaration itself not his declarations (has it’s a new
scope).

* Add test to ensure that nested class don’t resolve accidentally

* Resolve nested type usages within the same file

* Fix processDeclarations under a declaration block

They were previously ignored and so resolve failed to find declaration
declared under a declaration block of a ConditionalStatement

* Add test to ensure enum value can’t be resolved as a type

* Add test for fully qualified name struct to resolve in nested case

* Add test for fully qualified struct should not resolve a top level struct

* Introduce new GenericExpressionElementReference for reference elements

* Add a simple generic resolve algorithm

* Use the new GenericExpressionReference in ReferenceExpression

* Add processDeclaration to EnumDeclaration

* Use the generic resolve for expression reference

* #865 Uncomment an enum value resolve test

* Don’t process module declaration twice

Because Module declaration is marked as Declaration, it was previously
handled in both if statements.

* Search for imported symbol directly at the processDeclaration level

This makes the search more generic and available for all processors
implementations.

* Add search (processDeclaration) for implicitely imported object module

* Change Highlight annotator to use resolve

* Add test to resolve local imports

* Add tests for local foreach variable resolution

* Fix resolve for variable out of foreach scope

The variable lifetime introduced by the foreach is the foreach statement
itself, no more.

* Add test covering For loop introduced variable

* Fix resolve for a variable in the for loop condition refering the
declared variable in the for loop

* Add test about variable declaration initialization value referencing itself

* Fix resolve of value in variable declaration

* Fix an anonymous enum test

If it’s an anonymous enum, it don’t have a qualifier

* Fix declaration inside a declaration block not found

* Add test about the resolve with declaration inside a declaration block

* Fix resolve of elements inside declarationBlock and AttributeSpecifier

* Fix import tests when alias (rename)

The aliased reference should resolve to the alias declaration. The goto
will be handled by another mechanism. With this correct behavior a
rename is possible.

* Add tests for elements in selective import resolve

* Fix import resolution algorithm

* Add test about a resolve in an alias declaration resolving to another
module

* Remove invalid tests

Refering to the specs, theses tests were invalid

* Don’t warn if symbol is defined multiple times (for
polyVariantReferences)

It’s normal, that’s why it’s a PolyVariantReference.

* Introduce a new reference type for ImportBind and use it

* Remove all the previous implementation of Resolve

Now that the new system is in place, the old hacks can disapear.

* Remove commented block of code in scope processor impl

This code was about statement, so uneeded as statements don’t bring a
declaration (so don’t need a processDeclarations to be implemented).

* Add a processDeclaration to SingleImport

* Fix implementation of import in a specific case

* Comments some tests where implementation can’t easily cover properly

The current implementation limitations of import managment prevents
these tests to pass, so comment them for now.

* Fix scope module operator search

* Enable a test with scope module operator and imports

* Handle template search

* Fix undefined symbol in case of template type element

* Disable highlighting test

The resolve state is not able for now to Handle CTFE properly so disable
the test otherwise it fails.

* Update changelog

* Fix compilation

---------

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>