Add type comprehension support (#1022)
* Fix content of ReferenceExpression and FunctionCallExpression

* Remove non existent node type

* Remove TemplateMixinExpression node, this node does not exists anymore

* Unary expression: update elements inside based on new parsing

* Pragma: add note to remove expression from his name and remove
expression from his interfaces

Pragma is not an expression, D specs were updated to not show the
expression word anymore (as it is confusing). We should do the same
thing to remove the ambiguity.

* Add types support

This support is not complete and a lot of TODOs remains. But this is
enough to start improve the resolving algorithm

* Use type for improved resolve

* Add resolve test with aliased type

* Add resolve for local definition in ConditionalStatement

* fix(#587): Make PossiblyUndefinedSymbol not reporting error for compiler defined
symbol

* Add resolve for anonymous struct and union fields

* Add test for imported anonymous enum member

* Fix resolve for alias in some cases

Alias value can point to something else than type

* Fix resolve of auto assignment with name shadowing

* Fix caching when psi generation is modified

With this modification, a change in the script will invalidate caches
and so deleted nodes won’t reappear

---------

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>