Fix(#1028): add missing declaration to json module in plugin.xml
For Dub Module, we depends on Json to check the validity of the schema
(and so helps the user to determine if the dub.json is vadid and provide
autocompletion). But since Intelij 2024.3, json was moved into his own
module, so add it to allow the application to work correctly.