Harden the code block element reparsing
With some invalid code, the parser could stop eagerly and do not parse
everyting. Add a recovery code to ensure that the whole text is covered
by the builder and the IDE is happy.

This could be really improved (with better detection of errors and
recovening in the parser itself but it will be done in a dedicated PR of
better errors detections).