Fix(#1114): null pointer exception when creating a dub project
This part of the code was not updated and still used the old way to
access the dub binary. Fixed it by migrating to the new
dubBinaryProvider service. An improved way for the future, is to
generate ourself the assets and not rely on the Dub CLI (allows a
generation of asserts even if dub binary is missing).