Improve dub module creation UI (#1113)
* Change creation ui of Dlang Module and project for intellij

With this change the creation of module & project look like other
languages. An extension is added to allow dub to wire itself to the D
Language page (like for gradle and Maven in Java). This allow to have
dub being more decoupled to D core, and to have better UI and better UX.

* Move an extension declaration specific to intellij in idea.xml

* Move class specific to idea in idea module

And convert it to kotlin

* Remove unused class

This class was intended to be used for clion but was not used. And the
Project page will probably never be implemented and remain an idea
feature only.

* Remove custom step in module builder

* Update changelog

---------

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>