Handle fully qualified name references resolve
To do so, use the PsiPackage and PsiModule elements. These are an
abstraction that can be backed by directory or by file as previously.
The general behavior of go to for package and module do not change, but
this abstraction allow to properly handle references to a package during
the resolve.