Remove a lexer test (#1165)
This test covers a full D std file, it is useless to cover a such
complex file to just know how are handled tokens. It tends to just
accept whatever the lexer generate without actually check.

Lexer for D Language is simple, it is easy to covers the full accepted
range of tokens without having to cover a full complex file.

This kind of file is more useful for parser.

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>