Fix clion compatibility (#1155)
* Remove a dependency in java compiler

The "hack" to prevent build directory generation with sources files
copied in it do not seems to be needed anymore. Also it is incompatible
for clion.

* Fix plugin XML structure

declaring multiple times depends optional on the same module do not work
as expected and is marked as a warning. Actually this causes the plugin
to be incompatible to Clion, because the second optional will be
considered as mandatory.

---------

Co-authored-by: Etienne Brateau <etienne.brateau@gmail.com>